Main class being broken into smaller more functional classes
